+++ tests/api/test_text.py
